Made adjustments according to the feedback from my pull request
Used eslint to fix formatting issues and changed the directory structure for the components I was working on.
